Move artifact transform related codecs to their own package
